Pushed the node class into it's own class
